+++ FindYourFriends/App.js
@@ -8,9 +8,10 @@ import ChatScreen from "./screens/ChatScreen";
 import { ApolloClient } from "apollo-client";
 import { InMemoryCache } from "apollo-cache-inmemory";
 import { ApolloProvider } from "@apollo/react-hooks";
-import authLink from "./utils/links";
+import client from "./utils/ApolloClientProvider";
 import { createHttpLink } from "apollo-link-http";
 import { onError } from "apollo-link-error";
+import MakeAlert from "./utils/ErrorHandler";
 
 export default function App() {
   // USE SCREENS LIKE THIS
@@ -21,102 +22,7 @@ export default function App() {
     title: "An Error Occurred",
   });
 
-  const backendUri = "http://cdccc558.ngrok.io";
-  /**
-   * Put logic here, on how to handle errors.
-   * This handles all GraphQL errors.
-   * It was not my intention to put this here. I wanted it to be in utils/links.
-   * But I couldn't figure out how to get setError to utils/links.
-   */
-  const errorLink = onError(({ graphQLErrors, networkError }) => {
-    if (graphQLErrors) {
-      let errorMessage = {
-        ...error,
-      };
-      graphQLErrors.map((err) => {
-        const { message, locations, path } = err;
-        console.log(
-          `[GraphQL error]: 
-          Message: ${message}, 
-          Location: ${JSON.stringify(locations, null, 4)}, 
-          Path: ${path}, 
-          \nFull Error: ${JSON.stringify(err, null, 4)}\n\n`
-        );
-        switch (err.extensions.code) {
-          case "UNAUTHENTICATED":
-            // Unauthenticated Error from backend.
-            // Add logic here like "If not authenticated, send to login-page"
-            // And set errorMessage to user via setError,
-            // that they tried something that required login, but they weren't logged in
-            errorMessage = {
-              message: errorMessage.message.concat(message + "\n"),
-              title: "Unauthenticated.",
-            };
-          case "FORBIDDEN":
-            // ForbiddenError from backend.
-            // Should probably also send to login.
-            errorMessage = {
-              message: errorMessage.message.concat(message + "\n"),
-              title: "Unauthorized action.",
-            };
-          case "BAD_USER_INPUT":
-            errorMessage = {
-              message: errorMessage.message
-                .concat(`Following fields were wrong: 
-                ${err.extensions.exception.invalidArgs}
-                \n${message}\n`),
-              title: "Bad user input.",
-            };
-          default:
-            // Open Alert box with message.
-            errorMessage.message = errorMessage.message.concat(message + "\n");
-        }
-      });
-    }
-    if (networkError) console.log(`[Network error]: ${networkError}`);
-    setError({ ...errorMessage });
-  });
-  // the URI key is a string endpoint or function resolving to an endpoint -- will default to "/graphql" if not specified
-  const httpLink = createHttpLink({ uri: backendUri + "/graphql" });
-  /** httpLink and cache are requirements as of Apollo 2 https://www.apollographql.com/docs/react/api/apollo-client/#required-fields
-   * - Error link - https://www.apollographql.com/docs/link/links/error/
-   * - In Memory Cache - https://www.apollographql.com/docs/angular/basics/caching/
-   * Whenever Apollo Client fetches query results from your server,
-   * it automatically caches those results locally.
-   * This makes subsequent executions of the same query extremely fast.
-   */
-  const client = new ApolloClient({
-    uri: backendUri,
-    link: errorLink.concat(authLink.concat(httpLink)),
-    cache: new InMemoryCache(), // automatic caching
-  });
-
-  /**
-   * Errorhandling Alert.
-   * Tapping any button will fire the respective onPress callback and dismiss the alert.
-   * @param {string} error Text describing the error that occurred.
-   * @param {string} title The Title for the Error Alert. Default "An Error Occurred"
-   * @param {string} buttonText The text on the button. Default "OK"
-   */
-  const MyAlert = (error, title = "An Error Occurred", buttonText = "OK") =>
-    // static alert(title, message?, buttons?, options?)
-    Alert.alert(
-      title,
-      error,
-      [
-        {
-          text: buttonText,
-          onPress: () => {
-            console.log(JSON.stringify({ error }, null, 4));
-            setError({ message: null, title: "An Error Occurred" });
-            console.log(`${buttonText} pressed on Error Alert.`);
-          },
-        },
-      ],
-      { cancelable: false } // If false, you can only dismiss via the onPress. If True, you can dismiss by pressing outside the alert.
-    );
-
-  let content = <HomeScreen setTest={setTest} />;
+  let content = <HomeScreen setTest={setTest} setError={setError} />;
   if (test) {
     //content = <MapScreen test={test} />
     content = (
@@ -133,7 +39,7 @@ export default function App() {
   return (
     <ApolloProvider client={client}>
       <View style={styles.screen}>
-        {error.message && MyAlert(error.message, error.title)}
+        <MakeAlert error={error} setError={setError} />
         <Header title="Find Your Friends" />
         {content}
       </View>
